requirements.txt
streamlit
langchain
langchain-openai
pandas
beautifulsoup4
python-dotenv
google-generativeai 
langchain-google-genai
black
chromadb
faiss-cpu
PyPDF2